requirements.txt
absl-py==2.1.0
asgiref==3.8.1
astunparse==1.6.3
dj-database-url==2.2.0
dj-static==0.0.6
Django==5.1
django-environ==0.11.2
Flask==3.0.3
flickrapi==2.4.0
gunicorn==23.0.0
itsdangerous==2.2.0
joblib==1.4.2
keras==2.12.0
numpy==1.23.5
pillow==10.4.0
protobuf==3.20.3
psycopg2-binary==2.9.9
pyasn1-modules==0.2.8
pydrive==1.3.1
python-decouple==3.8
requests-toolbelt==1.0.0
scikit-learn==1.5.1
sqlparse==0.5.1
static3==0.7.0
tflite-runtime==2.14.0
threadpoolctl==3.5.0
webencodings==0.5.1
whitenoise==6.7.0